added download log
