+++ Cyberdrop Downloader.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 15.0.28307.1259
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "cyberdropDownloader", "cyberdropDownloader\cyberdropDownloader.csproj", "{1F6A8A0E-9AA8-4E46-814F-3634D794A446}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{119C2F46-056A-4903-BB98-023416C0AD3D}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
