+++ cyberdropDownloader/packages.config
@@ -1,7 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Downloader" version="1.8.0" targetFramework="net461" />
+  <package id="Downloader" version="2.0.7" targetFramework="net461" />
   <package id="HtmlAgilityPack" version="1.11.24" targetFramework="net461" />
   <package id="Syroot.Windows.IO.KnownFolders" version="1.2.3" targetFramework="net461" />
+  <package id="System.Buffers" version="4.4.0" targetFramework="net461" />
+  <package id="System.Memory" version="4.5.3" targetFramework="net461" />
+  <package id="System.Numerics.Vectors" version="4.4.0" targetFramework="net461" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.2" targetFramework="net461" />
   <package id="System.Security.Principal.Windows" version="4.7.0" targetFramework="net461" />
 </packages>
